Termino de sesion heroes

+++ src/app/heroes/pages/agregar/agregar.component.ts
@@ -1,6 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { ActivatedRoute, Router } from '@angular/router';
 import { switchMap } from 'rxjs';
+import { ConfirmarComponent } from '../../components/confirmar/confirmar.component';
 import { Heroe, Publisher } from '../../interfaces/heroes.interface';
 import { HeroesService } from '../../services/heroes.service';
 
@@ -40,7 +43,9 @@ export class AgregarComponent implements OnInit {
   constructor(
     private heroesService: HeroesService,
     private activatedRoute: ActivatedRoute,
-    private router: Router
+    private router: Router,
+    private snackBar: MatSnackBar,
+    private dialog: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -61,16 +66,32 @@ export class AgregarComponent implements OnInit {
     if (this.heroe.id) {
       this.heroesService
         .actHeroe(this.heroe)
-        .subscribe((heroe) => console.log('Actualizando', heroe));
+        .subscribe(() => this.mostrarSnackbar('Registro Actualizado'));
     } else {
       this.heroesService.getHeroe(this.heroe).subscribe((heroe) => {
         this.router.navigate(['/heroes/editar', heroe.id]);
+        this.mostrarSnackbar('Registro Creado');
       });
     }
   }
   deleteHeroe() {
-    this.heroesService.deleteHeroe(this.heroe.id!).subscribe((resp) => {
-      this.router.navigate(['/heroes']);
+    const dialog = this.dialog.open(ConfirmarComponent, {
+      width: '250px',
+      data: this.heroe,
+    });
+
+    dialog.afterClosed().subscribe((results) => {
+      if (results) {
+        this.heroesService.deleteHeroe(this.heroe.id!).subscribe((resp) => {
+          this.router.navigate(['/heroes']);
+        });
+      }
+    });
+  }
+
+  mostrarSnackbar(mensaje: string) {
+    this.snackBar.open(mensaje, 'ok!', {
+      duration: 2500,
     });
   }
 }
